rename inner preview
