+++ apps/src/sites/studio/pages/codeprojects_preview/show.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import InnerHTMLPreview2 from '@cdo/apps/codebridge/FilePreview/InnerHTMLPreview2';
+import InnerHTMLPreview from '@cdo/apps/codebridge/FilePreview/InnerHTMLPreview';
 
 window.React = require('react');
 window.ReactDOM = require('react-dom');
 
 document.addEventListener('DOMContentLoaded', () => {
   ReactDOM.render(
-    <InnerHTMLPreview2 />,
+    <InnerHTMLPreview />,
     document.getElementById('codeprojects-preview-container')
   );
 });
